feat(ui): move node ui into expandible rows
